refactor NFC files into the mmu/unit/nfc structure.   Adjust logging to use the mmu logger, separate the commands to individual command files.
